requirements.txt
google-generativeai
requests
pandas
pyyaml
python-dotenv
streamlit
plotly